Updated the Application
Based on my code review, I removed the Week Object and made the DTO for WeekResponse a Map<String, DayResponse>. I also removed some error handling in the Controller and Service that were a little overboard. Inside of the service layer I extracted the uri builder to create a helper method and modernized the Constructor for the MealPlannerService.